requirements.txt
flask==2.0.1
flask-login==0.5.0
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
authlib==0.15.5
requests==2.26.0
gunicorn==20.1.0
google-cloud-storage==1.42.0
google-cloud-secret-manager==2.7.0
psycopg2-binary==2.9.1
apscheduler==3.8.1
python-dotenv==0.19.0